Put Piecewise class back in closes #15833
