+++ framework/include/functions/PiecewiseConstant.h
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include "PiecewiseBase.h"
+#include "Piecewise.h"
 
 // Forward declarations
 class PiecewiseConstant;
@@ -21,7 +21,7 @@ InputParameters validParams<PiecewiseConstant>();
  * Function which provides a piecewise continuous constant interpolation
  * of a provided (x,y) point data set.
  */
-class PiecewiseConstant : public PiecewiseBase
+class PiecewiseConstant : public Piecewise
 {
 public:
   static InputParameters validParams();
